requirements.txt
Django>=4.2
djangorestframework
requests
beautifulsoup4
gunicorn
Pillow
python-dotenv
selenium
webdriver-manager
